+++ src/index.ts
@@ -1,2 +1,9 @@
 export {AngularSsrDistribution, AngularSsrDistributionProps} from './angular-ssr-distribution';
-export {invalidateCdnStep} from './invalidate-cdn-step';
+export {AngularStaticDistribution, AngularStaticDistributionProps} from './angular-static-distribution';
+export {
+    buildStaticViewerRequestCode,
+    readRedirectManifest,
+    FUNCTION_SIZE_LIMIT_BYTES,
+    StaticViewerRequestOptions,
+} from './static-viewer-request';
+export {invalidateCdnStep, InvalidatableDistribution} from './invalidate-cdn-step';